# Remove assembly for decoding trades.