@fedgiac 2021-01-28
Mmh. The gas increase is bad but not very bad. The resulting code looks much nicer.

I implemented a draft of some planned gas optimizations in #382 and it would save just an extra ~140gas/trade, bringing the total cost of implementing this PR to ~1425. Also, the same encoding change should also be implemented for encoded interactions, adding some extra cost for each interaction.

I'm unsure but maybe slightly in favor of implementing this PR, paying the extra gas cost to avoid all the complexity. Especially so if we plan to make the contract upgradeable, in which case it's fine to have higher gas costs at the beginning.