@nlordell 2021-01-28
> Also, the same encoding change should also be implemented for encoded interactions, adding some extra cost for each interaction.

I suspect for interactions the difference will be less pronounced It would be `12 + x + 29` extra 0s where `x ∈[-1, 31]` so `[40, 72]` zeros or `[160, 288]` extra calldata gas, so an order of magnitude less.