@fleupold 2021-02-01
Generally I'm a fan of simplicity and have been awaiting the stabilisation of v2 encoding for a long time, so I'd be happy to use it. On the other hand you have invested a lot of effort into writing the assembly based encoding and I feel it actually looks quite readable/nice.

Also at the current gas price we'd be saving about 15c per order, which can compound over time as you mentioned. How about we start the audit with the contract as is and get some feedback from Adam on how high he estimates the risk of leaving the assembly encoding in place?

Maybe this discussion is also another "hint" that we should leave upgradability of the logic as an option.