@nlordell 2021-02-01
My main argument for this change is one of the conclusions from the 0x post-mortem writen above:

> Do not over-optimize code by using assembly.

I think the amount of gas here is non-trivial, and means, essentially, that users will be required to pay an additional 15¢ per order in fees. However, if we do implement wrapper contracts that adjust the solution data, then this benefit will mostly be negated by the added encoding complexity.